requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
et-xmlfile==1.1.0
Flask==2.2.2
google-search-results==2.4.1
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.16
MarkupSafe==2.1.1
# numpy==1.23.5
openai==0.25.0
openpyxl==3.0.10
# pandas==1.5.1
# pandas-stubs==1.5.1.221024
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.44
tqdm==4.64.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
